Move interrupt and notify into gui plugin rather than wxmmain
